// SimERP/ClientApp/src/app/lists/shippedprice/model/shippedprice.ts
export class ShippedPrice {
  Id: number;
  MaDoGiaVc = '';
  DienGiai = '';
  GiaVc: number;
  GhiChu: string;
  CreatedDate?: Date;
  CreatedBy?: number;
  ModifyDate?: Date;
  ModifyBy?: number;
  IsActive = true;
  IsDelete = false;
  SearchString: string;
  SortOrder?: number;
  StoreId?: number;
  StoreName: string;
}